requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
clip==0.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
ftfy==6.3.1
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnxruntime==1.22.1
onnxslim==0.1.68
openai-clip==1.0.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
protobuf==6.32.1
psutil==7.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.15.3
six==1.17.0
supervision==0.26.1
sympy==1.14.0
tk==0.1.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
triton==3.4.0
typing_extensions==4.15.0
ultralytics==8.3.197
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.4
